RF_00.00.005 | Pipeline > Join to one class
